Extract service & middleware setup to extensions
Move DI and middleware configuration out of Program.cs into two new extension classes: ServiceExtensions (adds AddApplicationServices to register Steam client, caching, CORS, rate limiting, API-key validation, local storage, etc.) and MiddlewareExtensions (ConfigurePipeline to set up forwarded headers, Cloudflare header parsing, CORS, antiforgery, rate limiter and error handling). Program.cs now calls AddApplicationServices and ConfigurePipeline. Also bump JoyfulReaperLib package from 0.0.1 to 0.0.2. This cleans up startup and centralizes configuration logic.
